requirements.txt
asgiref==3.8.1
Django==5.1.2
django-ranged-response==0.2.0
django-simple-captcha==0.6.0
pillow==11.0.0
psycopg==3.2.3
psycopg2-binary==2.9.10
pytz==2024.2
sqlparse==0.5.1
typing_extensions==4.12.2
